+++ packages/sitecore-jss-nextjs/src/editing/editing-config-middleware.ts
@@ -26,10 +26,13 @@ export type EditingConfigMiddlewareConfig = {
  * provides configuration information to determine feature compatibility on Pages side.
  */
 export class EditingConfigMiddleware {
+  private enableEditingMetadata: boolean;
   /**
    * @param {EditingConfigMiddlewareConfig} [config] Editing configuration middleware config
    */
-  constructor(protected config: EditingConfigMiddlewareConfig) {}
+  constructor(protected config: EditingConfigMiddlewareConfig) {
+    this.enableEditingMetadata = config.enableEditingMetadata ?? true;
+  }
 
   /**
    * Gets the Next.js API route handler
@@ -55,7 +58,7 @@ export class EditingConfigMiddleware {
       ? this.config.components
       : Array.from(this.config.components.keys());
 
-    const editMode = this.config.enableEditingMetadata ? EditMode.Metadata : EditMode.Chromes;
+    const editMode = this.enableEditingMetadata ? EditMode.Metadata : EditMode.Chromes;
 
     return res.status(200).json({
       components,